+++ src/main/groovy/org/nuisto/ErrorCodes.groovy
@@ -1,6 +1,8 @@
 package org.nuisto
 
 class ErrorCodes {
+  public static final int Success = 0
   public static final int RulesNotProvided = -1
   public static final int ProvidedDirectoryDoesNotExist = -2
+  public static final int OptionsNotProvided = -3
 }